:recycle: api: Migrate cetification errors to context
