+++ api/lib/domain/usecases/retrieve-last-or-create-certification-course.js
@@ -1,5 +1,6 @@
 import bluebird from 'bluebird';
 
+import { SessionNotAccessible } from '../../../src/certification/session-management/domain/errors.js';
 import { ComplementaryCertificationCourse } from '../../../src/certification/session-management/domain/models/ComplementaryCertificationCourse.js';
 import { CertificationCourse } from '../../../src/certification/shared/domain/models/CertificationCourse.js';
 import { CertificationVersion } from '../../../src/certification/shared/domain/models/CertificationVersion.js';
@@ -10,7 +11,6 @@ import {
   CandidateNotAuthorizedToJoinSessionError,
   CandidateNotAuthorizedToResumeCertificationTestError,
   NotFoundError,
-  SessionNotAccessible,
   UnexpectedUserAccountError,
   UserNotAuthorizedToCertifyError,
 } from '../errors.js';
