+++ api/src/certification/session-management/domain/constants.js
